Create Site & Hosts Entry based on TFS BranchName
Powershell script, which when placed in the root of a branch in TFS, can be used to quickly configure a local developers machine. The script will attempt to derive the name of the TFS Branch from the Workspace mapping. If it is able to derive a Branch Name, it will create a Website and AppPool in IIS as well as an entry in the users HOSTS file. 

Given a Branch Name of "Project1" for example, 
  - an AppPool called "Project1" will be created.
  - a WebSite named "Project1" will be created.
  - The WebSite "Project1" will have a binding "project1.localhost.com"
  - a Loopback entry in the HOSTS file will be created for "project1.localhost.com"

Script uses the Carbon PS Module, which will be installed if it is not present.